move tests files and improve .dockerignore
