+++ docker/.dockerignore
@@ -1,1 +1,7 @@
 node_modules
+**/__mocks__/**
+coverage
+test_datsets
+docker
+figma
+**/__test__/**
